requirements.txt
requests
beautifulsoup4
openai
faiss-cpu
tiktoken
langchain
streamlit
langchain-community
langchain-openai
flask
python-dotenv
gunicorn

python-docx

tqdm
